Move x86 cell functions to their own module
Relocate the sse2 and avx cell functions to the cell module
Provide a code example for using Cell
