+++ src/x86.rs
@@ -18,6 +18,8 @@ pub use self::sse2::*;
 mod avx;
 pub use self::avx::*;
 
+pub mod cell;
+
 // Internal module for sealing SIMD traits.
 mod private {
     pub trait Sealed {}
