+++ src/x86/avx.rs
@@ -5,11 +5,9 @@ use core::arch::x86_64::{self as arch, __m128, __m128d, __m256, __m256d, __m256i
 use core::ptr;
 
 #[cfg(target_arch = "x86")]
-use crate::x86::{Is128BitsUnaligned, Is128CellUnaligned, Is256BitsUnaligned, Is256CellUnaligned};
+use crate::x86::{Is128BitsUnaligned, Is256BitsUnaligned};
 #[cfg(target_arch = "x86_64")]
-use crate::x86_64::{
-    Is128BitsUnaligned, Is128CellUnaligned, Is256BitsUnaligned, Is256CellUnaligned,
-};
+use crate::x86_64::{Is128BitsUnaligned, Is256BitsUnaligned};
 
 /// Broadcasts 128 bits from memory (composed of 2 packed double-precision
 /// (64-bit) floating-point elements) to all elements of the returned vector.
@@ -110,15 +108,6 @@ pub fn _mm256_loadu_si256<T: Is256BitsUnaligned>(mem_addr: &T) -> __m256i {
     unsafe { arch::_mm256_loadu_si256(ptr::from_ref(mem_addr).cast()) }
 }
 
-/// Loads 256-bits of integer data from memory into result.
-///
-/// [Intel's documentation](https://www.intel.com/content/www/us/en/docs/intrinsics-guide/index.html#text=_mm256_loadu_si256)
-#[inline]
-#[target_feature(enable = "avx")]
-pub fn _mm256_loadu_si256_cell<T: Is256CellUnaligned>(mem_addr: &T) -> __m256i {
-    unsafe { arch::_mm256_loadu_si256(ptr::from_ref(mem_addr).cast()) }
-}
-
 /// Loads two 128-bit values (composed of 4 packed single-precision (32-bit)
 /// floating-point elements) from memory, and combine them into a 256-bit
 /// value.
@@ -151,16 +140,6 @@ pub fn _mm256_loadu2_m128i<T: Is128BitsUnaligned>(hiaddr: &T, loaddr: &T) -> __m
     unsafe { arch::_mm256_loadu2_m128i(ptr::from_ref(hiaddr).cast(), ptr::from_ref(loaddr).cast()) }
 }
 
-/// Loads two 128-bit values (composed of integer data) from memory, and combine
-/// them into a 256-bit value.
-///
-/// [Intel's documentation](https://www.intel.com/content/www/us/en/docs/intrinsics-guide/index.html#text=_mm256_loadu2_m128i)
-#[inline]
-#[target_feature(enable = "avx")]
-pub fn _mm256_loadu2_m128i_cell<T: Is128CellUnaligned>(hiaddr: &T, loaddr: &T) -> __m256i {
-    unsafe { arch::_mm256_loadu2_m128i(ptr::from_ref(hiaddr).cast(), ptr::from_ref(loaddr).cast()) }
-}
-
 /// Stores 256-bits (composed of 4 packed double-precision (64-bit)
 /// floating-point elements) from `a` into memory.
 ///
@@ -189,16 +168,6 @@ pub fn _mm256_storeu_ps(mem_addr: &mut [f32; 8], a: __m256) {
 #[target_feature(enable = "avx")]
 pub fn _mm256_storeu_si256<T: Is256BitsUnaligned>(mem_addr: &mut T, a: __m256i) {
     unsafe { arch::_mm256_storeu_si256(ptr::from_mut(mem_addr).cast(), a) }
-}
-
-/// Stores 256-bits of integer data from `a` into memory.
-/// `mem_addr` does not need to be aligned on any particular boundary.
-///
-/// [Intel's documentation](https://www.intel.com/content/www/us/en/docs/intrinsics-guide/index.html#text=_mm256_storeu_si256)
-#[inline]
-#[target_feature(enable = "avx")]
-pub fn _mm256_storeu_si256_cell<T: Is256CellUnaligned>(mem_addr: &T, a: __m256i) {
-    unsafe { arch::_mm256_storeu_si256(ptr::from_ref(mem_addr).cast_mut().cast(), a) }
 }
 
 /// Stores the high and low 128-bit halves (each composed of 4 packed
@@ -234,22 +203,6 @@ pub fn _mm256_storeu2_m128i<T: Is128BitsUnaligned>(hiaddr: &mut T, loaddr: &mut 
         arch::_mm256_storeu2_m128i(
             ptr::from_mut(hiaddr).cast(),
             ptr::from_mut(loaddr).cast(),
-            a,
-        )
-    }
-}
-
-/// Stores the high and low 128-bit halves (each composed of integer data) from
-/// `a` into memory two different 128-bit locations.
-///
-/// [Intel's documentation](https://www.intel.com/content/www/us/en/docs/intrinsics-guide/index.html#text=_mm256_storeu2_m128i)
-#[inline]
-#[target_feature(enable = "avx")]
-pub fn _mm256_storeu2_m128i_cell<T: Is128CellUnaligned>(hiaddr: &T, loaddr: &T, a: __m256i) {
-    unsafe {
-        arch::_mm256_storeu2_m128i(
-            ptr::from_ref(hiaddr).cast_mut().cast(),
-            ptr::from_ref(loaddr).cast_mut().cast(),
             a,
         )
     }
@@ -534,34 +487,6 @@ mod tests {
 
             assert_eq!(hi, [30, 40]);
             assert_eq!(lo, [10, 20]);
-        }
-    }
-
-    #[test]
-    fn test_mm256_storeu2_m128i_cell() {
-        assert!(*CPU_HAS_AVX);
-
-        unsafe { test() }
-
-        #[target_feature(enable = "avx")]
-        fn test() {
-            let mut st_hi: [i64; 3] = [30, 40, 0];
-            let mut st_lo: [i64; 3] = [10, 20, 0];
-
-            let hi = core::cell::Cell::from_mut(&mut st_hi[..]).as_slice_of_cells();
-            let lo = core::cell::Cell::from_mut(&mut st_lo[..]).as_slice_of_cells();
-
-            let lhi: &[_; 2] = hi[0..2].try_into().unwrap();
-            let llo: &[_; 2] = lo[0..2].try_into().unwrap();
-
-            let shi: &[_; 2] = hi[1..3].try_into().unwrap();
-            let slo: &[_; 2] = lo[1..3].try_into().unwrap();
-
-            let a = super::_mm256_loadu2_m128i_cell(lhi, llo);
-            super::_mm256_storeu2_m128i_cell(shi, slo, a);
-
-            assert_eq!(st_hi, [30, 30, 40]);
-            assert_eq!(st_lo, [10, 10, 20]);
         }
     }
 
@@ -874,29 +799,6 @@ mod tests {
             super::_mm256_storeu_si256(&mut x, r);
 
             assert_eq!(x, [-1, -2, -3, -4]);
-        }
-    }
-
-    #[test]
-    fn test_mm256_loadu_si256_cell() {
-        assert!(*CPU_HAS_AVX);
-
-        unsafe { test() }
-
-        #[target_feature(enable = "avx")]
-        fn test() {
-            let mut x: [i16; 18] = core::array::from_fn(|i| i as i16);
-            let whole_cell = core::cell::Cell::from_mut(&mut x[..]);
-
-            let in_cell: &[_; 16] = whole_cell.as_slice_of_cells()[..16].try_into().unwrap();
-            let mm256 = super::_mm256_loadu_si256_cell(in_cell);
-
-            let out_cell: &[_; 16] = whole_cell.as_slice_of_cells()[2..].try_into().unwrap();
-            super::_mm256_storeu_si256_cell(out_cell, mm256);
-
-            let y: [i16; 16] = core::array::from_fn(|i| i as i16);
-            // We copied it over into this area of the underlying storage
-            assert_eq!(y, x[2..]);
         }
     }
 }
