+++ src/x86/sse2.rs
@@ -5,9 +5,9 @@ use core::arch::x86_64::{self as arch, __m128d, __m128i};
 use core::ptr;
 
 #[cfg(target_arch = "x86")]
-use crate::x86::{Is128BitsUnaligned, Is128CellUnaligned};
+use crate::x86::Is128BitsUnaligned;
 #[cfg(target_arch = "x86_64")]
-use crate::x86_64::{Is128BitsUnaligned, Is128CellUnaligned};
+use crate::x86_64::Is128BitsUnaligned;
 
 /// Loads a double-precision (64-bit) floating-point element from memory
 /// into both elements of returned vector.
@@ -59,15 +59,6 @@ pub fn _mm_loadl_epi64<T: Is128BitsUnaligned>(mem_addr: &T) -> __m128i {
     unsafe { arch::_mm_loadl_epi64(ptr::from_ref(mem_addr).cast()) }
 }
 
-/// Loads a 64-bit integer from memory into first element of returned vector.
-///
-/// [Intel's documentation](https://www.intel.com/content/www/us/en/docs/intrinsics-guide/index.html#text=_mm_loadl_epi64)
-#[inline]
-#[target_feature(enable = "sse2")]
-pub fn _mm_loadl_epi64_cell<T: Is128CellUnaligned>(mem_addr: &T) -> __m128i {
-    unsafe { arch::_mm_loadl_epi64(ptr::from_ref(mem_addr).cast()) }
-}
-
 /// Loads a double-precision value into the low-order bits of a 128-bit
 /// vector of `[2 x double]`. The high-order bits are copied from the
 /// high-order bits of the first operand.
@@ -96,15 +87,6 @@ pub fn _mm_loadu_pd(mem_addr: &[f64; 2]) -> __m128d {
 #[target_feature(enable = "sse2")]
 pub fn _mm_loadu_si128<T: Is128BitsUnaligned>(mem_addr: &T) -> __m128i {
     unsafe { arch::_mm_loadu_si128(ptr::from_ref(mem_addr).cast()) }
-}
-
-/// Loads 128-bits of integer data from memory into a new vector.
-///
-/// [Intel's documentation](https://www.intel.com/content/www/us/en/docs/intrinsics-guide/index.html#text=_mm_loadu_si128)
-#[inline]
-#[target_feature(enable = "sse2")]
-pub fn _mm_loadu_si128_cell<T: Is128CellUnaligned>(mem_addr: &T) -> __m128i {
-    unsafe { arch::_mm_loadu_si128(ptr::from_ref(mem_addr).cast_mut().cast()) }
 }
 
 /// Loads unaligned 16-bits of integer data from memory into new vector.
@@ -163,15 +145,6 @@ pub fn _mm_storel_epi64<T: Is128BitsUnaligned>(mem_addr: &mut T, a: __m128i) {
     unsafe { arch::_mm_storel_epi64(ptr::from_mut(mem_addr).cast(), a) }
 }
 
-/// Stores the lower 64-bit integer `a` to a memory location.
-///
-/// [Intel's documentation](https://www.intel.com/content/www/us/en/docs/intrinsics-guide/index.html#text=_mm_storel_epi64)
-#[inline]
-#[target_feature(enable = "sse2")]
-pub fn _mm_storel_epi64_cell<T: Is128CellUnaligned>(mem_addr: &T, a: __m128i) {
-    unsafe { arch::_mm_storel_epi64(ptr::from_ref(mem_addr).cast_mut().cast(), a) }
-}
-
 /// Stores the lower 64 bits of a 128-bit vector of `[2 x double]` to a
 /// memory location.
 ///
@@ -199,15 +172,6 @@ pub fn _mm_storeu_pd(mem_addr: &mut [f64; 2], a: __m128d) {
 #[target_feature(enable = "sse2")]
 pub fn _mm_storeu_si128<T: Is128BitsUnaligned>(mem_addr: &mut T, a: __m128i) {
     unsafe { arch::_mm_storeu_si128(ptr::from_mut(mem_addr).cast(), a) }
-}
-
-/// Stores 128-bits of integer data from `a` into memory.
-///
-/// [Intel's documentation](https://www.intel.com/content/www/us/en/docs/intrinsics-guide/index.html#text=_mm_storeu_si128)
-#[inline]
-#[target_feature(enable = "sse2")]
-pub fn _mm_storeu_si128_cell<T: Is128CellUnaligned>(mem_addr: &T, a: __m128i) {
-    unsafe { arch::_mm_storeu_si128(ptr::from_ref(mem_addr).cast_mut().cast(), a) }
 }
 
 /// Store 16-bit integer from the first element of `a` into memory.
@@ -309,21 +273,6 @@ mod tests {
         #[target_feature(enable = "sse2")]
         fn test(a: &[i64; 2]) {
             let r = super::_mm_loadl_epi64(a);
-            let target = arch::_mm_set_epi64x(0, 20);
-
-            assert_eq_m128i(r, target)
-        }
-    }
-
-    #[test]
-    fn test_mm_loadl_epi64_cell() {
-        let mut a = [20, 25];
-        unsafe { test(&mut a) }
-
-        #[target_feature(enable = "sse2")]
-        fn test(a: &mut [i64; 2]) {
-            let val = core::cell::Cell::from_mut(a);
-            let r = super::_mm_loadl_epi64_cell(val);
             let target = arch::_mm_set_epi64x(0, 20);
 
             assert_eq_m128i(r, target)
@@ -440,40 +389,6 @@ mod tests {
             super::_mm_storel_epi64(&mut x, a);
 
             assert_eq!(x[0], i64::MAX);
-        }
-    }
-
-    #[test]
-    fn test_mm_storel_epi64_cell() {
-        unsafe { test() }
-
-        #[target_feature(enable = "sse2")]
-        fn test() {
-            let a = arch::_mm_set_epi64x(i64::MIN, i64::MAX);
-            let mut x = [0; 2];
-            let val = core::cell::Cell::from_mut(&mut x);
-            super::_mm_storel_epi64_cell(val, a);
-
-            assert_eq!(x[0], i64::MAX);
-        }
-    }
-
-    #[test]
-    fn test_mm_storeu_si128_epi32_cell() {
-        unsafe { test() }
-
-        #[target_feature(enable = "sse2")]
-        fn test() {
-            let mut a = [0u32, 1, 2, 3, 4];
-            let val = core::cell::Cell::from_mut(&mut a[..]).as_slice_of_cells();
-
-            let load: &[_; 4] = val[..4].try_into().unwrap();
-            let store: &[_; 4] = val[1..].try_into().unwrap();
-
-            let r = super::_mm_loadu_si128_cell(load);
-            super::_mm_storeu_si128_cell(store, r);
-
-            assert_eq!(a, [0, 0, 1, 2, 3]);
         }
     }
 
@@ -663,25 +578,6 @@ mod tests {
             let target = arch::_mm_set_epi64x(-2, -1);
 
             assert_eq_m128i(r, target);
-        }
-    }
-
-    #[test]
-    fn test_mm_loadu_si128_i64_cell() {
-        let mut a = [-1, -2, 0];
-        unsafe { test(&mut a) }
-
-        #[target_feature(enable = "sse2")]
-        fn test(a: &mut [i64; 3]) {
-            let val = core::cell::Cell::from_mut(&mut a[..]).as_slice_of_cells();
-
-            let load: &[_; 2] = val[..2].try_into().unwrap();
-            let store: &[_; 2] = val[1..].try_into().unwrap();
-
-            let r = super::_mm_loadu_si128_cell(load);
-            super::_mm_storeu_si128_cell(store, r);
-
-            assert_eq!(*a, [-1, -1, -2]);
         }
     }
